# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "description": "A human approval or rejection of a plan.",
        "properties": {
          "approver": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Approver",
            "type": "string"
          },
          "change_id": {
            "title": "Change Id",
            "type": "string"
          },
          "comment": {
            "default": "",
            "maxLength": 500,
            "title": "Comment",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          }
        },
        "required": [
          "change_id",
          "approver",
          "decision"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "CheckAnswer": {
        "additionalProperties": false,
        "description": "A learner's answer to one knowledge check.",
        "properties": {
          "choice": {
            "maximum": 9.0,
            "minimum": 0.0,
            "title": "Choice",
            "type": "integer"
          }
        },
        "required": [
          "choice"
        ],
        "title": "CheckAnswer",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Control plane for the offline-first Fabric + Foundry reference implementation.",
    "title": "Fabric Foundry Integration Accelerator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals": {
      "post": {
        "description": "Record a human approval or rejection.",
        "operationId": "approve_api_v1_approvals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Approval"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "changes"
        ]
      }
    },
    "/api/v1/audit/{correlation_id}": {
      "get": {
        "description": "Redacted audit records for a correlation ID.",
        "operationId": "get_audit_api_v1_audit__correlation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "correlation_id",
            "required": true,
            "schema": {
              "title": "Correlation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditRecord"
                  },
                  "title": "Response Get Audit Api V1 Audit  Correlation Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/v1/education/labs/{lab_id}": {
      "get": {
        "description": "One lab with its eleven stages.",
        "operationId": "get_lab_api_v1_education_labs__lab_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lab_id",
            "required": true,
            "schema": {
              "title": "Lab Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lab"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lab",
        "tags": [
          "education"
        ]
      }
    },
    "/api/v1/education/lessons": {
      "get": {
        "description": "Lesson summaries, optionally filtered by area or pattern.",
        "operationId": "list_lessons_api_v1_education_lessons_get",
        "parameters": [
          {
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area"
            }
          },
          {
            "in": "query",
            "name": "pattern_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pattern Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LessonSummary"
                  },
                  "title": "Response List Lessons Api V1 Education Lessons Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lessons",
        "tags": [
          "education"
        ]
      }
    },
    "/api/v1/education/lessons/{lesson_id}": {
      "get": {
        "description": "One lesson with all five levels and its knowledge checks (answers withheld).",
        "operationId": "get_lesson_api_v1_education_lessons__lesson_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "title": "Lesson Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lesson",
        "tags": [
          "education"
        ]
      }
    },
    "/api/v1/education/lessons/{lesson_id}/checks/{check_id}": {
      "post": {
        "description": "Grade one knowledge-check answer.",
        "operationId": "answer_check_api_v1_education_lessons__lesson_id__checks__check_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "lesson_id",
            "required": true,
            "schema": {
              "title": "Lesson Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "check_id",
            "required": true,
            "schema": {
              "title": "Check Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAnswer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckGrade"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Check",
        "tags": [
          "education"
        ]
      }
    },
...
    "/api/v1/education/views/{view_id}": {
      "get": {
        "description": "One architecture view and its layout: nodes, edges, zones, bands, build steps and traces.",
        "operationId": "get_view_api_v1_education_views__view_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "view_id",
            "required": true,
            "schema": {
              "title": "View Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get View",
        "tags": [
          "education"
        ]
      }
    },
    "/api/v1/education/views/{view_id}/drawio": {
      "get": {
        "description": "The view as a draw.io file (same content as docs/architecture/diagrams).",
        "operationId": "get_view_drawio_api_v1_education_views__view_id__drawio_get",
        "parameters": [
          {
            "in": "path",
            "name": "view_id",
            "required": true,
            "schema": {
              "title": "View Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get View Drawio",
        "tags": [
          "education"
        ]
      }
    },
    "/api/v1/education/views/{view_id}/runtime": {
      "get": {
        "description": "The live state of every runtime-bound node in a view.",
        "operationId": "get_view_runtime_api_v1_education_views__view_id__runtime_get",
        "parameters": [
          {
            "in": "path",
            "name": "view_id",
            "required": true,
            "schema": {
              "title": "View Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewRuntime"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get View Runtime",
        "tags": [
          "education"
        ]
      }
    },
    "/api/v1/evaluations/run": {
      "post": {
        "description": "Compare observed (or local) measure values with the expected baseline.",
        "operationId": "run_evaluation_api_v1_evaluations_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEnvelope_object_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Evaluation",
        "tags": [
          "evaluation"
        ]
      }
    },
    "/api/v1/fabric/change": {
      "post": {
        "description": "Execute an approved change. LIVE changes are never redirected to LOCAL.",
        "operationId": "execute_change_api_v1_fabric_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEnvelope_object_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Change",
        "tags": [
          "changes"
        ]
      }
    },
    "/api/v1/fabric/read": {
      "post": {
        "description": "Typed, read-only Fabric operations through the provider router.",
        "operationId": "fabric_read_api_v1_fabric_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ListWorkspaces"
                  },
                  {
                    "$ref": "#/components/schemas/ListItems"
                  },
                  {
                    "$ref": "#/components/schemas/ListTables"
                  },
                  {
                    "$ref": "#/components/schemas/ReadTable"
                  },
                  {
                    "$ref": "#/components/schemas/GetSemanticModel"
                  },
                  {
                    "$ref": "#/components/schemas/EvaluateMeasures"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionEnvelope_object_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fabric Read",
        "tags": [
          "fabric"
        ]
      }
    },
...
    "/api/v1/guides/{guide_id}": {
      "get": {
        "description": "One Use-Case Guide.",
        "operationId": "get_guide_api_v1_guides__guide_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "guide_id",
            "required": true,
            "schema": {
              "title": "Guide Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UseCaseGuide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Guide",
        "tags": [
          "guides"
        ]
      }
    },
    "/api/v1/guides/{guide_id}/steps/{step_id}": {
      "get": {
        "description": "One guide step.",
        "operationId": "get_guide_step_api_v1_guides__guide_id__steps__step_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "guide_id",
            "required": true,
            "schema": {
              "title": "Guide Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "step_id",
            "required": true,
            "schema": {
              "title": "Step Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuideStep"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Guide Step",
        "tags": [
          "guides"
        ]
      }
    },
...
  }
}